backend/auth_project/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
black==24.4.2
    # via -r requirements.in
build==1.2.1
    # via pip-tools
cffi==1.16.0
    # via cryptography
click==8.1.7
    # via
    #   black
    #   pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
cryptography==42.0.8
    # via jwt
django==5.0.6
    # via
    #   -r requirements.in
    #   djangorestframework
    #   djangorestframework-simplejwt
django-restframework==0.0.1
    # via -r requirements.in
djangorestframework==3.15.1
    # via djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
jwt==1.3.1
    # via -r requirements.in
mypy-extensions==1.0.0
    # via black
packaging==24.1
    # via
    #   black
    #   build
pathspec==0.12.1
    # via black
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.2.2
    # via black
psycopg2-binary==2.9.9
    # via -r requirements.in
pycparser==2.22
    # via cffi
pyjwt==2.8.0
    # via djangorestframework-simplejwt
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
pytz==2024.1
    # via -r requirements.in
sqlparse==0.5.0
    # via django
tzdata==2024.1
    # via django
wheel==0.43.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
